refactor: refactor recs dir (#1012)
